api-python/pyproject.toml
[tool.poetry]
name = "api-python"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.0"
gunicorn = "^19.9"
grpcio = "^1.19"
grpcio-tools = "^1.19"
voluptuous = "^0.11.5"
python-etcd = "^0.4.5"

[tool.poetry.dev-dependencies]
mypy = "^0.670.0"
pylint = "^2.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
